# docs/mobile/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/public/test-delete-section/{id}": {
      "get": {
        "operationId": "PublicController_testDeleteSection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Public"
        ]
      }
    },
...
    "/public/courses": {
      "get": {
        "operationId": "PublicController_getCourses",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Category slug",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "grade",
            "required": false,
            "in": "query",
            "description": "Grade/level slug",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "newest | rating | popular",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List published courses for public browsing",
        "tags": [
          "Public"
        ]
      }
    },
    "/public/courses/{slug}": {
      "get": {
        "operationId": "PublicController_getCourse",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a single published course by slug",
        "tags": [
          "Public"
        ]
      }
    },
    "/public/teachers": {
      "get": {
        "operationId": "PublicController_getTeachers",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subjectId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List approved teachers for public browsing",
        "tags": [
          "Public"
        ]
      }
    },
    "/public/teachers/{id}": {
      "get": {
        "operationId": "PublicController_getTeacher",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get teacher public profile with their courses",
        "tags": [
          "Public"
        ]
      }
    },
...
  },
  "info": {
    "title": "Masarak API",
    "description": "The Masarak Enterprise E-Learning API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}